requirements.txt
matplotlib==3.10.8
numba==0.64.0
numba_progress==1.1.3
numpy==2.4.2
Pillow==12.1.1
scipy==1.17.1
sympy==1.13.3
